timetable: Rename file names
